backend/trace-chat/requirements.txt
fastapi>=0.104.1
uvicorn>=0.24.0
python-dotenv>=1.0.0
crewai>=0.1.0
openai>=1.3.0
langchain>=0.0.350 